requirements.txt
<<<<<<< HEAD
discord.py==2.4.0
aiosqlite==0.20.0
=======
discord.py==2.4.0
aiosqlite==0.20.0
>>>>>>> 2d71b26ac8256f9e449e430bea05a0cd25c05fe2
